+++ CHANGELOG.md
@@ -6,6 +6,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.5.0] - Sep 11, 2021
+
+### Added
+
+* Add [Extractors](/docs/extractors)
+* Add ``network_mode`` support for docker containers
+* Add additional tests (extractors + network_mode)
+
+### Changed
+
+* Changed the log output format for docker containers
+* Changed exception handling of AttributeException in OsCommandPlugin
+* Small bug fixes
+
+### Removed
+
+* Removed six dependency (closes #4)
+* Removed short version for most options
+
+
 ## [1.4.2] - June 11, 2021
 
 ### Changed